Move EntityLinking to one common package
