+++ src/main.py
@@ -2,7 +2,7 @@ import argparse
 from pathlib import Path
 
 from src.Embeddings.embeddings_factory import EmbeddingsFactory
-from src.EntityLinkingsConfiguration.entity_linkings_factory import EntityLinkingsFactory
+from src.EntityLinking.entity_linkings_factory import EntityLinkingsFactory
 from src.TaskConfiguration import TaskCategoryCollectionFactory
